+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Improve the `NotAuthorizedError` message to include the policy class.
+- Improve the `NotAuthorizedError` message to include the policy class. (#812)
 Furthermore, in the case where the record passed is a class instead of an instance, the class name is given.
 
 ## 2.3.2 (2024-05-08)